Seperated hyperparameter optimization to a separate file
